# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Reddit Research AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ReportStatusProperty = typeof ReportStatusProperty[keyof typeof ReportStatusProperty];


export const ReportStatusProperty = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * Structured report JSON when completed
 * @nullable
 */
export type ReportResult = { [key: string]: unknown } | null;

export interface Report {
  id: number;
  userId: string;
  keyword: string;
  /** @nullable */
  subreddit?: string | null;
  /** @nullable */
  timeRange?: string | null;
  /** @nullable */
  maxPosts?: number | null;
  /** @nullable */
  maxComments?: number | null;
  status: ReportStatusProperty;
  /**
     * Progress percentage 0-100
     * @nullable
     */
  progress?: number | null;
  /** @nullable */
  progressMessage?: string | null;
  /**
     * Structured report JSON when completed
     * @nullable
     */
  result?: ReportResult;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  postsAnalyzed?: number | null;
  /** @nullable */
  commentsAnalyzed?: number | null;
  /** @nullable */
  aiProvider?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ReportInputTimeRange = typeof ReportInputTimeRange[keyof typeof ReportInputTimeRange];


export const ReportInputTimeRange = {
  day: 'day',
  week: 'week',
  month: 'month',
  year: 'year',
  all: 'all',
} as const;

export interface ReportInput {
  /** @minLength 1 */
  keyword: string;
  subreddit?: string;
  timeRange?: ReportInputTimeRange;
  /**
     * @minimum 1
     * @maximum 100
     */
  maxPosts?: number;
  /**
     * @minimum 10
     * @maximum 500
     */
  maxComments?: number;
  /** ID of the saved API key to use for LLM inference */
  apiKeyId: number;
}

export type ReportStatusStatus = typeof ReportStatusStatus[keyof typeof ReportStatusStatus];


export const ReportStatusStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ReportStatus {
  id: number;
  status: ReportStatusStatus;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  progressMessage?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export type ApiKeyProvider = typeof ApiKeyProvider[keyof typeof ApiKeyProvider];


export const ApiKeyProvider = {
  openai: 'openai',
  anthropic: 'anthropic',
  gemini: 'gemini',
  perplexity: 'perplexity',
  openrouter: 'openrouter',
  groq: 'groq',
  mistral: 'mistral',
  deepseek: 'deepseek',
  cohere: 'cohere',
  xai: 'xai',
} as const;

export interface ApiKey {
  id: number;
  userId: string;
  provider: ApiKeyProvider;
  name: string;
  /** Last 4 characters of the key for display */
  keyPreview: string;
  isActive: boolean;
  createdAt: string;
}

export type ApiKeyInputProvider = typeof ApiKeyInputProvider[keyof typeof ApiKeyInputProvider];


export const ApiKeyInputProvider = {
  openai: 'openai',
  anthropic: 'anthropic',
  gemini: 'gemini',
  perplexity: 'perplexity',
  openrouter: 'openrouter',
  groq: 'groq',
  mistral: 'mistral',
  deepseek: 'deepseek',
  cohere: 'cohere',
  xai: 'xai',
} as const;

export interface ApiKeyInput {
  provider: ApiKeyInputProvider;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  key: string;
}

export type ApiKeyValidateInputProvider = typeof ApiKeyValidateInputProvider[keyof typeof ApiKeyValidateInputProvider];


export const ApiKeyValidateInputProvider = {
  openai: 'openai',
  anthropic: 'anthropic',
  gemini: 'gemini',
  perplexity: 'perplexity',
  openrouter: 'openrouter',
  groq: 'groq',
  mistral: 'mistral',
  deepseek: 'deepseek',
  cohere: 'cohere',
  xai: 'xai',
} as const;

export interface ApiKeyValidateInput {
  provider: ApiKeyValidateInputProvider;
  /** @minLength 1 */
  key: string;
}

export interface ApiKeyValidateResult {
  valid: boolean;
  /** @nullable */
  message?: string | null;
}

export interface KeywordCount {
  keyword: string;
  count: number;
}

export interface DashboardSummary {
  totalReports: number;
  completedReports: number;
  pendingReports: number;
  failedReports: number;
  recentReports: Report[];
  topKeywords: KeywordCount[];
}