# lib/api-zod/src/generated/types/apiKey.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Reddit Research AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKeyProvider } from './apiKeyProvider';

export interface ApiKey {
  id: number;
  userId: string;
  provider: ApiKeyProvider;
  name: string;
  /** Last 4 characters of the key for display */
  keyPreview: string;
  isActive: boolean;
  createdAt: Date;
}