# lib/api-zod/src/generated/types/apiKeyInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Reddit Research AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKeyInputProvider } from './apiKeyInputProvider';

export interface ApiKeyInput {
  provider: ApiKeyInputProvider;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  key: string;
}