# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Reddit Research AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KeywordCount } from './keywordCount';
import type { Report } from './report';

export interface DashboardSummary {
  totalReports: number;
  completedReports: number;
  pendingReports: number;
  failedReports: number;
  recentReports: Report[];
  topKeywords: KeywordCount[];
}