# lib/api-zod/src/generated/types/keywordCount.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Reddit Research AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KeywordCount {
  keyword: string;
  count: number;
}