# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Reddit Research AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportInputTimeRange } from './reportInputTimeRange';

export interface ReportInput {
  /** @minLength 1 */
  keyword: string;
  subreddit?: string;
  timeRange?: ReportInputTimeRange;
  /**
     * @minimum 1
     * @maximum 100
     */
  maxPosts?: number;
  /**
     * @minimum 10
     * @maximum 500
     */
  maxComments?: number;
  /** ID of the saved API key to use for LLM inference */
  apiKeyId: number;
}